requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
amqp==5.2.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
billiard==4.2.1
bleach==6.2.0
celery==5.3.6
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.2
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.2
django-appconf==1.1.0
django-celery-beat==2.8.0
django-compressor==4.5.1
django-cors-headers==4.7.0
django-crontab==0.7.1
django-timezone-field==7.1
django-vite==3.1.0
django_csp==3.8
djangorestframework==3.15.2
dotenv==0.9.9
frozenlist==1.6.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
kombu==5.3.4
lxml==5.3.0
multidict==6.4.3
oauthlib==3.2.2
orjson==3.10.16
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
rcssmin==1.1.2
redis==4.6.0
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.2
ruff==0.11.6
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
stripe==12.0.1
tomli==2.2.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
yarl==1.20.0
pylint-django==2.5.3
